Update gallery creation
