+++ jsx/UcButton.tsx
@@ -2,32 +2,36 @@ import React from "react/index"
 import { Button } from "@wordpress/components";
 import UcUploader from "./UcUploader";
 import config from './uc-config';
+import FileInfoResponse from "./FileInfoResponse";
 
 const wp = (window as any).wp;
 const { __ } = wp.i18n;
 const uploader = new UcUploader(config.config);
 
-const upload = async () => {
+const upload = () => {
     try {
-        const data = await uploader.upload();
         const block = wp.data.select('core/block-editor').getSelectedBlock();
         if (block === null) return false;
 
-        if (block.name === 'core/gallery') {
-            const newImage = {
-                fullUrl: data.cdnUrl,
-                url: data.cdnUrl,
+        uploader.upload().then((data: FileInfoResponse) => {
+            if (block.name === 'core/gallery') {
+                console.log(data.attach_id, data)
+                block.attributes.images.push({
+                    fullUrl: data.cdnUrl,
+                    url: data.cdnUrl,
+                    // id: `"${data.attach_id}"`
+                });
+                // block.attributes.ids.push(data.attach_id);
             }
-            block.attributes.images.push(newImage)
-        }
 
-        if (block.name === 'core/image') {
-            block.attributes.url = data.cdnUrl;
-            block.attributes.alt = data.name;
-        }
+            if (block.name === 'core/image') {
+                block.attributes.url = data.cdnUrl;
+                block.attributes.alt = data.name;
+            }
 
-        wp.data.dispatch('core/block-editor').clearSelectedBlock();
-        wp.data.dispatch('core/block-editor').replaceBlock(block.clientId, block);
+            wp.data.dispatch('core/block-editor').clearSelectedBlock();
+            wp.data.dispatch('core/block-editor').replaceBlock(block.clientId, block);
+        });
     } catch (err) {
         document.querySelectorAll('div.uploadcare-loading-screen').forEach(el => {
             el.classList.add('uploadcare-hidden')
